BREAKING CHANGE: move count to new endpoint
